+++ python/GafferUITest/NumericPlugValueWidgetTest.py
@@ -51,12 +51,12 @@ class NumericPlugValueWidgetTest( GafferUITest.TestCase ) :
 
 		w = GafferUI.NumericPlugValueWidget( n["i"] )
 		self.assertTrue( w.getPlug().isSame( n["i"] ) )
-		GafferUITest.PlugValueWidgetTest.waitForUpdate( w )
+		self.waitForPlugValueWidgetUpdate( w )
 		self.assertTrue( isinstance( w.numericWidget().getValue(), int ) )
 
 		w.setPlug( n["f"] )
 		self.assertTrue( w.getPlug().isSame( n["f"] ) )
-		GafferUITest.PlugValueWidgetTest.waitForUpdate( w )
+		self.waitForPlugValueWidgetUpdate( w )
 		self.assertTrue( isinstance( w.numericWidget().getValue(), float ) )
 
 		w = GafferUI.NumericPlugValueWidget( plugs = [] )
@@ -65,7 +65,7 @@ class NumericPlugValueWidgetTest( GafferUITest.TestCase ) :
 
 		w.setPlug( n["f"] )
 		self.assertTrue( w.getPlug().isSame( n["f"] ) )
-		GafferUITest.PlugValueWidgetTest.waitForUpdate( w )
+		self.waitForPlugValueWidgetUpdate( w )
 		self.assertTrue( isinstance( w.numericWidget().getValue(), float ) )
 		self.assertEqual( w.numericWidget().getEditable(), True )
 
@@ -80,11 +80,11 @@ class NumericPlugValueWidgetTest( GafferUITest.TestCase ) :
 
 		n["user"]["i1"].setValue( 2 )
 		n["user"]["i2"].setValue( 2 )
-		GafferUITest.PlugValueWidgetTest.waitForUpdate( w )
+		self.waitForPlugValueWidgetUpdate( w )
 		self.assertEqual( w.numericWidget().getText(), "2" )
 
 		n["user"]["i1"].setValue( 1 )
-		GafferUITest.PlugValueWidgetTest.waitForUpdate( w )
+		self.waitForPlugValueWidgetUpdate( w )
 		self.assertEqual( w.numericWidget().getText(), "" )
 		self.assertEqual( w.numericWidget()._qtWidget().placeholderText(), "---" )
 
@@ -104,7 +104,7 @@ class NumericPlugValueWidgetTest( GafferUITest.TestCase ) :
 		n["user"]["i2"].setValue( 2 )
 
 		w = GafferUI.NumericPlugValueWidget( n["user"]["i1"] )
-		GafferUITest.PlugValueWidgetTest.waitForUpdate( w )
+		self.waitForPlugValueWidgetUpdate( w )
 		self.assertEqual( w.numericWidget().getText(), "1" )
 		self.assertEqual( w.numericWidget()._qtWidget().placeholderText(), "" )
 
@@ -113,7 +113,7 @@ class NumericPlugValueWidgetTest( GafferUITest.TestCase ) :
 
 		self.assertEqual( n["user"]["i1"].getValue(), 1 )
 		self.assertEqual( n["user"]["i2"].getValue(), 2 )
-		GafferUITest.PlugValueWidgetTest.waitForUpdate( w )
+		self.waitForPlugValueWidgetUpdate( w )
 		self.assertEqual( w.numericWidget().getText(), "" )
 		self.assertEqual( w.numericWidget()._qtWidget().placeholderText(), "---" )
 
@@ -126,7 +126,7 @@ class NumericPlugValueWidgetTest( GafferUITest.TestCase ) :
 		n["user"]["i2"].setValue( 2 )
 
 		w = GafferUI.NumericPlugValueWidget( n["user"]["i1"] )
-		GafferUITest.PlugValueWidgetTest.waitForUpdate( w )
+		self.waitForPlugValueWidgetUpdate( w )
 		self.assertEqual( w.numericWidget().getValue(), 1 )
 
 		w.numericWidget().setText( "" )
@@ -135,7 +135,7 @@ class NumericPlugValueWidgetTest( GafferUITest.TestCase ) :
 		self.assertEqual( n["user"]["i1"].getValue(), 1 )
 
 		w = GafferUI.NumericPlugValueWidget( n["user"].children() )
-		GafferUITest.PlugValueWidgetTest.waitForUpdate( w )
+		self.waitForPlugValueWidgetUpdate( w )
 		self.assertEqual( w.numericWidget().getText(), "" )
 
 		w.numericWidget()._qtWidget().editingFinished.emit()