+++ Changes.md
@@ -28,7 +28,14 @@ Fixes
 API
 ---
 
-- GafferUITest.TestCase : Added `uiThreadCallHandler` member. This should be used by any tests which need to handle calls to the UI thread.
+- GafferUITest.TestCase :
+  - Added `uiThreadCallHandler` member. This should be used by any tests which need to handle calls to the UI thread.
+  - Added `waitForPlugValueWidgetUpdate()` method.
+
+Breaking Changes
+----------------
+
+- PlugValueWidgetTest : Removed `waitForUpdate()` method. Use `GafferUITest.TestCase.waitForPlugValueWidgetUpdate()` instead.
 
 Build
 -----